==> EMF/US/ml_emf/bin/classifier/1_Deepstarr_dataprep_ATAC.diffbind.deseq2-oneCL.ipynb <==
20240306

SarahFong

Take output from DiffBind deseq2 formatting

          $HOME/EMF/US/bin/ATAC/1_prep_deseq2_ATAC.ipynb 
          
Prepare data: 
1. Clean diffbind deseq2 conc data (remove bad coordinates)
4. Format dataframe for LegNet, multitask, deepstarr multitask
5. Split into training and test set - randomly hold out 2 chromosomes. 
6. Add fold num to training set

from Bio.SeqIO.FastaIO import SimpleFastaParser
import config_readwrite as crw
import matplotlib.pyplot as plt
import numpy as np
import os, sys
import pandas as pd
import pybedtools as pbt

from scipy import stats

import seaborn as sns

import warnings
warnings.filterwarnings('ignore')

import plot_params as pp
pp.fonts()

config_name =os.path.join(os.getcwd(), "config.multi.ini")
config, cfn = crw.read(config_name)

# functions

def setPrefixRules(prefix):
    """ rules for processing datasets"""
    if "class.all" in prefix:
        QUANTILE_FILTER=False
        MIN_CONC_FILTER=False
        JOINT_ACCESSIBLE=True
        CLASS_LABEL = True

    elif "class.nojoint" in prefix:
        QUANTILE_FILTER=False
        MIN_CONC_FILTER=True
        JOINT_ACCESSIBLE=False
        CLASS_LABEL = True

    elif "reg.all" in PREFIX:
        QUANTILE_FILTER=False
        MIN_CONC_FILTER=True
        JOINT_ACCESSIBLE=True
        CLASS_LABEL = False

    elif "reg.nojoint" in prefix:
        QUANTILE_FILTER=False
        MIN_CONC_FILTER=True
        JOINT_ACCESSIBLE=False
        CLASS_LABEL = False
    else:
        print('need to add rules for', prefix)
        
    return QUANTILE_FILTER, MIN_CONC_FILTER, JOINT_ACCESSIBLE, CLASS_LABEL


## make chromosome list

def chrList():
    """return  list of chromosomes"""
    
    chrs = []
    
    for n in np.arange(1,23):
        chrs.append(f"chr{n}")
    
    # add sex chromosomes
    chrs.append("chrX")
    chrs.append("chrY")
    
    return chrs

## write fa

def writeFa(heldout_df, heldout_fa):
    with open(heldout_fa, "w") as writer:
        for row in heldout_df.iterrows():
            seqid, seq=row[1][:2]
            writer.write(f">{seqid}\n{seq}\n")

## train test split on chromosome

def splitTrainTestVal(df, val_chr_list=None, test_chr_list=None):
    """randomly sample and hold out 2 chromosomes for testing, validation"""
    
    cols =['coor.type',"seq"]
    
    # randomly sample test chromosomes (n=2)
    if test_chr_list is None:
        print('randomly sampling chromosomes for test')


        # randomly sample
        test_chr_list = list(np.random.choice(chrs, 2))
   
    # randomly sample validation chromosome (n=1)
    if val_chr_list is None:
        print('randomly sampling chromosomes for val')

        # remove test chromosomes
        for chr_ in test_chr_list:
            chrs.remove(chr_)

        # randomly sample
        val_chr_list = list(np.random.choice(chrs, 1))

    # separate held out chromosomes from  training chromosomes
    test = df.loc[df["#chr"].isin(test_chr_list)].copy()
    val = df.loc[df["#chr"].isin(val_chr_list)].copy()
    train = df.loc[(~df["#chr"].isin(test_chr_list))&
                  (~df["#chr"].isin(val_chr_list))].copy()

    return train, val, test


## filters

### need a deseq2 filter. 

def minReadDepthFilter(table, MIN_CONC):

    print("before read depth filter:", table.shape)
    table = table.loc[(table["ctrl"] > MIN_CONC) |
                      (table["US"] > MIN_CONC)].copy()
    print("after:", table.shape)

    return table


def quantileFilter(table, quantile=0.99):
    """upper quantile filter for read count values"""

    print("before quantile filter:", table.shape)

    # quantiles
    ctrl_thresh, us_thresh = table[["ctrl", "US"]].quantile(quantile)

    # filter table
    table = table.loc[(table["ctrl"] < ctrl_thresh) &
                      (table["US"] < us_thresh)]

    print("after:", table.shape)

    return table


def jointAccessibleFilter(table):
    """remove joint accessible regions, scramble new dataframe"""

    print("before joint_accessible filter:", table.shape)

    # filter table
    ctrl_only = table.loc[(table["ctrl"] > 0) &
                          (table["US"] == 0)].copy()

    US_only = table.loc[(table["ctrl"] == 0) &
                        (table["US"] > 0)].copy()

    # combine us and ctrl
    # shuffle dataframe
    table = pd.concat([ctrl_only, US_only]).sample(
        frac=1).reset_index(drop=True)

    print("after:", table.shape)

    return table

# binarize


def classLabel(table):
    """binarize read count column"""
    cols_to_label = ['ctrl', 'US']
    for col in cols_to_label:
        table[col] = table[col].apply(lambda x: 0 if x<MIN_CONC else 1)

    return table

# label deseq DiffAcc peaks
def diffAccLabel(table, diff_bed):
    """ Annotate elements that overlap desseq2 differentially accessible peak calls
        Return dataframe of overlapping elements. 
    """
    table_bed = pbt.BedTool.from_dataframe(table[table.columns[:4]])
    deseq_bed = pbt.BedTool(diff_bed)

    table_deseq = table_bed.intersect(deseq_bed, wao=True)

    table_deseq = pd.read_csv(table_deseq.fn, sep='\t', header=None)
    table_deseq.columns=['#chr', "start_trim", "end_trim", "type", 
                         "#chr_deseq", "start_deseq", "end_deseq", 
                         "deseq_id", "diffAcc"
                        ]

    #annotate diffAcc overlap as 1 or 0
    table_deseq["diffAcc"] = table_deseq["diffAcc"].apply(lambda x: 1 if x>0 else 0) 

    # merge annotations with original table
    table = pd.merge(table, table_deseq[["type","diffAcc"]]).drop_duplicates()
    
    return table

# scramble df


def dfShuffle(df):
    """scramble dataframe"""
    return df.sample(frac=1).reset_index(drop=True)

def directionFilter(table):
    """filter for positive directional changes in activity"""

    table = table.loc[table["US_DIF"]<0] #where US_DIF = ctrl - US, and US_dif <0 means US has more activity than control
    
    return table

# Main

## load data

CL = 'hob'
PREFIX = "class.all.2tasks"

PATH = "/wynton/group/ahituv/data/US-MPRA/ATAC-seq/Diffbind_results"
DATA_PATH = f"/wynton/home/ahituv/fongsl/EMF/US/ml_emf/data/deepstarr/deseq2/{CL}/{PREFIX}"

TEST_CHR = ["chr8"]
VAL_CHR = "chr12"


# peak information
PEAK_SIZE = 270
MIN_CONC = 3
DIRECTION_FILTER = False  # keep only positive changes? Negative changes? 

# genome information
HG38= "/wynton/group/ahituv/data/dna/hg38/hg38.chrom.sizes"
FA_HG38="/wynton/group/ahituv/data/dna/hg38/hg38.fa"


if "class" in PREFIX:
    cols=['coor.type', "seq", "diffAcc", "fold_num"]
else:
    cols = ['coor.type', "seq", "ctrl", "US", "US_DIF", "fold_num"]
    
fa_cols = ['coor.type', "seq"]


if os.path.exists(DATA_PATH) is False:
    if os.path.dirname(DATA_PATH) is False:
        os.mkdir(os.path.dirname(DATA_PATH))
    os.mkdir(DATA_PATH)

## base files

# files

# DESEQ2 information
DIFF = f'./{CL}_deseq2.csv'
NODIFF = f'./{CL}_deseq2-nodiff.csv'
DIFF_BED = "./" + DIFF.strip(".csv") + ".bed"

# write training, test files
FULL = f"{CL}_deseq2-nondiff.trimmed.full.csv"


## base config

section = f"{CL}-ATAC-DESEQ2"
crw.check(config, section)

config[section]["path"] = PATH

config[section]["nondiff_bind_results"] = "%(path)s/" + NODIFF
config[section]["diff_bind_results"] = "%(path)s/" + DIFF
config[section]["diff_bind_results_bed"] = "%(path)s/" + DIFF_BED


## deepstarr files

# file inputs for training, testing
## x
TRAIN_FA = f"{PREFIX}.Sequences_Train.fa"
VAL_FA = f"{PREFIX}.Sequences_Val.fa"
TEST_FA = f"{PREFIX}.Sequences_Test.fa"

## y
TRAIN_TARGET = f"{PREFIX}.Sequences_activity_Train.txt"
VAL_TARGET = f"{PREFIX}.Sequences_activity_Val.txt"
TEST_TARGET = f"{PREFIX}.Sequences_activity_Test.txt"

## rules

## deepstarr config 

section = f"{CL}.atac.deseq2.deepstarr"
crw.check(config, section)

# deepstar config
config[section]["data_path"] = DATA_PATH
config[section]["held_out_chr"] = ",".join(TEST_CHR)
config[section]["val_chr"] = VAL_CHR

## deepstarr+prefix config

# rules for data
QUANTILE_FILTER, MIN_CONC_FILTER, JOINT_ACCESSIBLE, CLASS_LABEL = setPrefixRules(PREFIX)


# deepstarr prefix config
section = f"{CL}.atac.deseq2.deepstarr.{PREFIX}"
crw.check(config, section)

config[section]["train_fa"] = TRAIN_FA
config[section]["val_fa"] = VAL_FA
config[section]["test_fa"] = TEST_FA

config[section]["train_target"] = TRAIN_TARGET
config[section]["tval_target"] = VAL_TARGET
config[section]["test_target"] = TEST_TARGET


config[section]["filter_MIN_CONC"] = str(MIN_CONC_FILTER)
if MIN_CONC_FILTER is True:
    config[section]["MIN_CONC"] = str(MIN_CONC)
    
config[section]["filter_quantile"] = str(QUANTILE_FILTER)
config[section]["filter_jointaccessible"] = str(JOINT_ACCESSIBLE)
config[section]["classlabel"] = str(CLASS_LABEL)

crw.write(config, cfn)

os.chdir(PATH)
table = pd.read_csv(FULL, sep='\t').drop_duplicates()

table["coor.type"] = table["type"] + "|" + table["seq.id"]

if "Conc_Ultrasound_dif" not in list(table):
    table["Conc_Ultrasound_dif"] = table["Conc_Control"] - \
        table["Conc_Ultrasound"]

table.rename(columns={"Conc_Ultrasound": "US",
                      "Conc_Control": "ctrl",
                      "Conc_Ultrasound_dif": "US_DIF"}, inplace=True)


table.head()

table.shape

##  min read depth filter

## mean read depth filter
if MIN_CONC_FILTER is True:
    table = minReadDepthFilter(table, MIN_CONC) 

## quantile filter
if QUANTILE_FILTER is True:

    table = quantileFilter(table, quantile=0.99)

## joint accessible filter
if JOINT_ACCESSIBLE is False:
    table = jointAccessibleFilter(table)

## apply class label
if CLASS_LABEL is True:
    table = diffAccLabel(table, DIFF_BED)
    print(table.groupby(['ctrl', "US", "US_DIF", "diffAcc"])['#chr'].count())
    
if DIRECTION_FILTER is True:
    table = directionFilter(table)

table

## train on all atac peaks

os.chdir(DATA_PATH)

table = dfShuffle(table)  # shuffle the table before splitting

train, val, test = splitTrainTestVal(table, val_chr_list=[VAL_CHR], test_chr_list=TEST_CHR)

writeFa(test[fa_cols], TEST_FA)
writeFa(train[fa_cols], TRAIN_FA)
writeFa(val[fa_cols], VAL_FA)

if "reg" in PREFIX:
    
    cols = ["coor.type",
        "US",
        "ctrl", 
       # "US_DIF"
       ]
elif "class" in PREFIX:
    cols=["coor.type", "diffAcc"]
    
test[cols].to_csv(TEST_TARGET, sep='\t', index=False)
train[cols].to_csv(TRAIN_TARGET, sep='\t', index=False)
val[cols].to_csv(VAL_TARGET, sep='\t', index=False)

val.shape, train.shape, test.shape

train.loc[train["diffAcc"]>0] # ctrl - US <0 means US has more activity than control

train.loc[train["diffAcc"]>0].shape[0]/train.shape[0]

val.loc[val["diffAcc"]>0].shape[0]/val.shape[0]

test.loc[test["diffAcc"]>0].shape[0]/test.shape[0]

1024/train.shape[0]

64/train.shape[0]
