==> EMF/US/ml_emf/bin/classifier/2_launch_deepstarr.ipynb <==
# launch deepstarr

import os, sys

cl = "pos"#'hepg2'
peak_calls = "deseq2"  # or count_matrix
prefix = f"class.all" #"reg.all.8tasks.25epoch" # class.all class.nojoint reg.all reg.nojoint

balanced_window = ".balanced_window"
if balanced_window is not None:
    data_path = os.path.join("/wynton/home/ahituv/fongsl/EMF/US/ml_emf/data/deepstarr/", 
                         peak_calls, cl, prefix+balanced_window)
else:
    data_path = os.path.join("/wynton/home/ahituv/fongsl/EMF/US/ml_emf/data/deepstarr/", 
                         peak_calls, cl, prefix)

#data_path = "/wynton/home/ahituv/fongsl/MPRA/agarwal_2023/"
#prefix = f"class.{cl}.balanced" #"reg.all.8tasks.25epoch" # class.all class.nojoint reg.all reg.nojoint

if "reg" in prefix:
    prediction_task = "reg"
else:
    prediction_task = "class"
    
#standard_scaling = "False"
print(data_path)

def launchCmd(prefix, data_path, prediction_task, terminal="qsub"):
    if terminal == "python":
        cmd = ["python",
               "deepstarr.py",
               prefix,
               data_path,
               prediction_task,
               #standard_scaling, 

               ]

    elif terminal == "qsub":
        cmd = ["qsub",
               "deepstarr.cpu.sh",
               prefix,
               data_path,
               prediction_task,
               #standard_scaling, 

               ]
    elif terminal == "qsub-q":
        cmd = ["qsub -q gpu.q",
               "deepstarr.gpu.sh",
               prefix,
               data_path,
               prediction_task,
               #standard_scaling, 

               ]

    return " ".join(cmd)

cmd = launchCmd(prefix, data_path, prediction_task, terminal="qsub")
cmd

os.system(cmd)

import tensorflow as tf
