==> EMF/US/ml_emf/bin/classifier/2_launch_deepstarr.py.ipynb <==
import os, sys

prefix = "class.all"
data_path = "/wynton/home/ahituv/fongsl/EMF/US/ml_emf/data/deepstarr"
prediction_task = "class" # "reg"

def launchCmd(prefix, data_path, terminal="qsub"):
    if terminal == "python":
        cmd = ["python",
               "deepstarr.py",
               prefix,
               data_path
               ]

    elif terminal == "qsub":
        cmd = ["qsub",
               "deepstarr.cpu.sh",
               prefix,
               data_path
               ]
    return " ".join(cmd)

launchCmd(prefix, data_path, terminal="qsub")
