==> EMF/US/ml_emf/bin/classifier/3_performance_eval_pred.ipynb <==
import os, sys
import numpy as np
import pandas as pd

import h5py

import tensorflow as tf
from tensorflow import keras

PREFIX = "hepg2.us.ctrl.reg"

MODEL_NAME="DeepSTARR_ATAC"
DATA_PATH = "/wynton/home/ahituv/fongsl/EMF/US/ml_emf/data/deepstarr"

os.chdir(DATA_PATH)
# write model
#model_json = main_model.to_json()
#with open(f'Model_' + model_name + f'.{prefix}.json', "w") as json_file:
 #   json_file.write(model_json)
    
# load weights
m = os.path.join(DATA_PATH, f'Model_{MODEL_NAME}.{PREFIX}.json')

%%bash
head /wynton/home/ahituv/fongsl/EMF/US/ml_emf/data/deepstarr/Model_DeepSTARR_ATAC.hepg2.us.ctrl.reg.h5

tf.keras.models.load_model(m)

list("a"*2)
